Move OrderDetails out of common models to Portfolio.
