+++ broker-impl/src/main/java/com/redelastic/stocktrader/broker/impl/BrokerServiceImpl.java
@@ -16,7 +16,7 @@ import com.redelastic.stocktrader.broker.impl.order.OrderEntity;
 import com.redelastic.stocktrader.broker.impl.order.OrderEvent;
 import com.redelastic.stocktrader.broker.impl.order.OrderRepository;
 import com.redelastic.stocktrader.broker.impl.quote.QuoteService;
-import com.redelastic.stocktrader.order.OrderId;
+import com.redelastic.stocktrader.OrderId;
 import com.redelastic.stocktrader.portfolio.api.OrderPlaced;
 import com.redelastic.stocktrader.portfolio.api.PortfolioService;
 import org.slf4j.Logger;
