+++ broker-impl/src/main/java/com/redelastic/stocktrader/broker/impl/order/OrderCommand.java
@@ -6,8 +6,8 @@ import com.lightbend.lagom.serialization.Jsonable;
 import com.redelastic.stocktrader.PortfolioId;
 import com.redelastic.stocktrader.broker.api.OrderResult;
 import com.redelastic.stocktrader.broker.api.OrderStatus;
-import com.redelastic.stocktrader.order.Order;
-import com.redelastic.stocktrader.order.OrderDetails;
+import com.redelastic.stocktrader.portfolio.api.order.Order;
+import com.redelastic.stocktrader.portfolio.api.order.OrderDetails;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
