+++ broker-impl/src/main/java/com/redelastic/stocktrader/broker/impl/order/OrderRepositoryImpl.java
@@ -8,8 +8,8 @@ import com.lightbend.lagom.javadsl.persistence.PersistentEntityRegistry;
 import com.redelastic.stocktrader.broker.api.OrderResult;
 import com.redelastic.stocktrader.broker.api.Trade;
 import com.redelastic.stocktrader.broker.impl.trade.TradeService;
-import com.redelastic.stocktrader.order.Order;
-import com.redelastic.stocktrader.order.OrderId;
+import com.redelastic.stocktrader.portfolio.api.order.Order;
+import com.redelastic.stocktrader.OrderId;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
