+++ broker-impl/src/main/java/com/redelastic/stocktrader/broker/impl/trade/TradeService.java
@@ -1,7 +1,7 @@
 package com.redelastic.stocktrader.broker.impl.trade;
 
 import com.redelastic.stocktrader.broker.api.OrderResult;
-import com.redelastic.stocktrader.order.Order;
+import com.redelastic.stocktrader.portfolio.api.order.Order;
 
 import java.util.concurrent.CompletionStage;
