+++ broker-impl/src/main/java/com/redelastic/stocktrader/broker/impl/trade/TradeServiceImpl.java
@@ -4,9 +4,9 @@ import com.redelastic.CSHelper;
 import com.redelastic.stocktrader.broker.api.OrderResult;
 import com.redelastic.stocktrader.broker.api.Trade;
 import com.redelastic.stocktrader.broker.impl.quote.QuoteService;
-import com.redelastic.stocktrader.order.Order;
-import com.redelastic.stocktrader.order.OrderDetails;
-import com.redelastic.stocktrader.order.OrderType;
+import com.redelastic.stocktrader.portfolio.api.order.Order;
+import com.redelastic.stocktrader.portfolio.api.order.OrderDetails;
+import com.redelastic.stocktrader.portfolio.api.order.OrderType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
