+++ commonModels/src/main/java/com/redelastic/stocktrader/TransferId.java
@@ -1,7 +1,9 @@
 package com.redelastic.stocktrader;
 
+import com.fasterxml.jackson.annotation.JsonValue;
 import com.lightbend.lagom.javadsl.api.deser.PathParamSerializer;
 import com.lightbend.lagom.javadsl.api.deser.PathParamSerializers;
+import lombok.Getter;
 import lombok.NonNull;
 import lombok.Value;
 import org.pcollections.ConsPStack;
@@ -11,6 +13,7 @@ import java.util.UUID;
 
 @Value
 public class TransferId {
+
     @NonNull String id;
 
     public static TransferId newId() { return new TransferId(UUID.randomUUID().toString()); }
