+++ portfolio-api/src/main/java/com/redelastic/stocktrader/portfolio/api/OrderCompleted.java
@@ -4,7 +4,7 @@ package com.redelastic.stocktrader.portfolio.api;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.redelastic.stocktrader.PortfolioId;
-import com.redelastic.stocktrader.order.OrderId;
+import com.redelastic.stocktrader.OrderId;
 import lombok.EqualsAndHashCode;
 import lombok.NonNull;
 import lombok.Value;
