+++ portfolio-api/src/main/java/com/redelastic/stocktrader/portfolio/api/OrderPlaced.java
@@ -1,8 +1,8 @@
 package com.redelastic.stocktrader.portfolio.api;
 
 import com.redelastic.stocktrader.PortfolioId;
-import com.redelastic.stocktrader.order.OrderDetails;
-import com.redelastic.stocktrader.order.OrderId;
+import com.redelastic.stocktrader.portfolio.api.order.OrderDetails;
+import com.redelastic.stocktrader.OrderId;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NonNull;
