+++ portfolio-api/src/main/java/com/redelastic/stocktrader/portfolio/api/order/OrderDetails.java
@@ -1,5 +1,6 @@
-package com.redelastic.stocktrader.order;
+package com.redelastic.stocktrader.portfolio.api.order;
 
+import com.redelastic.stocktrader.TradeType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NonNull;
