+++ portfolio-impl/src/main/java/com/redelastic/stocktrader/portfolio/impl/PortfolioCommand.java
@@ -5,8 +5,8 @@ import com.lightbend.lagom.javadsl.persistence.PersistentEntity.ReplyType;
 import com.lightbend.lagom.serialization.Jsonable;
 import com.redelastic.stocktrader.broker.api.OrderResult;
 import com.redelastic.stocktrader.broker.api.Trade;
-import com.redelastic.stocktrader.order.OrderDetails;
-import com.redelastic.stocktrader.order.OrderId;
+import com.redelastic.stocktrader.portfolio.api.order.OrderDetails;
+import com.redelastic.stocktrader.OrderId;
 import lombok.Builder;
 import lombok.NonNull;
 import lombok.Value;
