+++ portfolio-impl/src/test/java/com/redelastic/stocktrader/portfolio/impl/PortfolioModelEntityTest.java
@@ -4,11 +4,11 @@ import akka.actor.ActorSystem;
 import akka.testkit.javadsl.TestKit;
 import com.lightbend.lagom.javadsl.testkit.PersistentEntityTestDriver;
 import com.redelastic.stocktrader.PortfolioId;
+import com.redelastic.stocktrader.TradeType;
 import com.redelastic.stocktrader.broker.api.Trade;
-import com.redelastic.stocktrader.order.OrderId;
-import com.redelastic.stocktrader.order.OrderType;
-import com.redelastic.stocktrader.order.OrderDetails;
-import com.redelastic.stocktrader.order.TradeType;
+import com.redelastic.stocktrader.OrderId;
+import com.redelastic.stocktrader.portfolio.api.order.OrderType;
+import com.redelastic.stocktrader.portfolio.api.order.OrderDetails;
 import com.redelastic.stocktrader.portfolio.impl.PortfolioCommand.Open;
 import com.redelastic.stocktrader.portfolio.impl.PortfolioCommand.PlaceOrder;
 import lombok.val;
