+++ portfolio-impl/src/test/java/com/redelastic/stocktrader/portfolio/impl/PortfolioServiceImplTest.java
@@ -8,12 +8,12 @@ import com.lightbend.lagom.javadsl.api.ServiceCall;
 import com.lightbend.lagom.javadsl.api.broker.Topic;
 import com.lightbend.lagom.javadsl.testkit.ProducerStub;
 import com.lightbend.lagom.javadsl.testkit.ProducerStubFactory;
+import com.redelastic.stocktrader.TradeType;
 import com.redelastic.stocktrader.TransferId;
 import com.redelastic.stocktrader.broker.api.*;
-import com.redelastic.stocktrader.order.OrderDetails;
-import com.redelastic.stocktrader.order.OrderId;
-import com.redelastic.stocktrader.order.OrderType;
-import com.redelastic.stocktrader.order.TradeType;
+import com.redelastic.stocktrader.portfolio.api.order.OrderDetails;
+import com.redelastic.stocktrader.OrderId;
+import com.redelastic.stocktrader.portfolio.api.order.OrderType;
 import com.redelastic.stocktrader.portfolio.api.*;
 import com.redelastic.stocktrader.wiretransfer.api.Transfer;
 import com.redelastic.stocktrader.wiretransfer.api.TransferRequest;
